FEAT(cart): separate cart component in Header
Jira: [[FE]separate cart component in Header](https://cozyhometeam.atlassian.net/browse/CH-625)
