+++ src/shared-components/Header/CartIcon/CartIcon.scss
@@ -1,0 +1,20 @@
+.cart-icon {
+    display: block;
+    position: relative;
+    &__counter {
+        position: absolute;
+        top: -5px;
+        right: -8px;
+        font-size: 10px;
+        font-weight: 500;
+        position: absolute;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        color: white;
+        width: 16px;
+        height: 16px;
+        background-color: #C57100;
+        border-radius: 100%;
+    }
+}
